+++ Tests/BugsnagTests.m
@@ -30,8 +30,7 @@
  * A boilerplate helper method to setup Bugsnag
  */
 -(void)setUpBugsnagWillCallNotify:(bool)willNotify {
-    NSError *error;
-    BugsnagConfiguration *configuration = [[BugsnagConfiguration alloc] initWithApiKey:DUMMY_APIKEY_32CHAR_1 error:&error];
+    BugsnagConfiguration *configuration = [[BugsnagConfiguration alloc] initWithApiKey:DUMMY_APIKEY_32CHAR_1];
     if (willNotify) {
         [configuration addOnSendBlock:^bool(BugsnagEvent * _Nonnull event) { return false; }];
     }
@@ -125,8 +124,7 @@
  *       or mocking is required to isolate and test the session pausing semantics.
  */
 -(void)testBugsnagPauseSession {
-    NSError *error;
-    BugsnagConfiguration *configuration = [[BugsnagConfiguration alloc] initWithApiKey:DUMMY_APIKEY_32CHAR_1 error:&error];
+    BugsnagConfiguration *configuration = [[BugsnagConfiguration alloc] initWithApiKey:DUMMY_APIKEY_32CHAR_1];
     [configuration addOnSendBlock:^bool(BugsnagEvent * _Nonnull event) { return false; }];
 
     [Bugsnag startBugsnagWithConfiguration:configuration];
@@ -142,8 +140,7 @@
     // Allow for checks inside blocks that may (potentially) be run asynchronously
     __block XCTestExpectation *expectation1 = [self expectationWithDescription:@"Localized metadata changes"];
     
-    NSError *error;
-    BugsnagConfiguration *configuration = [[BugsnagConfiguration alloc] initWithApiKey:DUMMY_APIKEY_32CHAR_1 error:&error];
+    BugsnagConfiguration *configuration = [[BugsnagConfiguration alloc] initWithApiKey:DUMMY_APIKEY_32CHAR_1];
     [configuration setContext:@"firstContext"];
     [configuration addOnSendBlock:^bool(BugsnagEvent * _Nonnull event) { return false; }];
     
@@ -225,8 +222,7 @@
     __block XCTestExpectation *expectation2 = [self expectationWithDescription:@"Remove On Session Block 2"];
     expectation2.inverted = YES;
     
-    NSError *error;
-    BugsnagConfiguration *configuration = [[BugsnagConfiguration alloc] initWithApiKey:DUMMY_APIKEY_32CHAR_1 error:&error];
+    BugsnagConfiguration *configuration = [[BugsnagConfiguration alloc] initWithApiKey:DUMMY_APIKEY_32CHAR_1];
 
     // non-sending bugsnag
     [configuration addOnSendBlock:^bool(BugsnagEvent * _Nonnull event) { return false; }];
@@ -262,8 +258,7 @@
     __block XCTestExpectation *expectation2 = [self expectationWithDescription:@"Remove On Session Block 3X"];
     expectation2.inverted = YES;
     
-    NSError *error;
-    BugsnagConfiguration *configuration = [[BugsnagConfiguration alloc] initWithApiKey:DUMMY_APIKEY_32CHAR_1 error:&error];
+    BugsnagConfiguration *configuration = [[BugsnagConfiguration alloc] initWithApiKey:DUMMY_APIKEY_32CHAR_1];
     configuration.autoTrackSessions = NO;
     
     // non-sending bugsnag
@@ -308,8 +303,7 @@
     __block int called = 0; // A counter
     
     // Prevent sending events
-    NSError *error;
-    BugsnagConfiguration *configuration = [[BugsnagConfiguration alloc] initWithApiKey:DUMMY_APIKEY_32CHAR_1 error:&error];
+    BugsnagConfiguration *configuration = [[BugsnagConfiguration alloc] initWithApiKey:DUMMY_APIKEY_32CHAR_1];
     // We'll not be able to use the onSend -> false route to fail calls to notify()
     [configuration setEndpointsForNotify:@"http://not.valid.bugsnag/not/an/endpoint"
                                 sessions:@"http://not.valid.bugsnag/not/an/endpoint"];