+++ CHANGELOG.md
@@ -29,10 +29,11 @@ Bugsnag Notifiers on other platforms.
   from the default convenience `init()` to ensure that `apiKey` has a value set.  The `apiKey`
   must now be a correctly formatted one to be accepted.
 
-  * Swift: `try BugsnagConfiguration(_ apiKey)`
-  * Objective C: `[[BugsnagConfiguration alloc] initWithApiKey:error:]`
+  * Swift: `BugsnagConfiguration(_ apiKey)`
+  * Objective C: `[[BugsnagConfiguration alloc] initWithApiKey:]`
 
   [#446](https://github.com/bugsnag/bugsnag-cocoa/pull/446)
+  [#492](https://github.com/bugsnag/bugsnag-cocoa/pull/492)
 
 * Support editing breadcrumbs within an Event as objects. Breadcrumbs can now be
   inspected and modified from callbacks, for example: